+++ databases/constants.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from typing import List, cast, get_args
 
 from lib import pyright
-from .types import SupportedDatabase
+from ._types import SupportedDatabase
 from .utils import DatabaseConfig, DatabaseFeature
